utrace_set_events: consolidate "setting _UTRACE_DEATH_EVENTS" checks
utrace_set_events() checks the "setting _UTRACE_DEATH_EVENTS" case twice,
and it is not immediately obvious why the first check is needed, and why
it is not racy (we are checking exit_state without tasklist). The code is
correct, but looks confusing.

The first check covers the case when target->utrace_flags already has
_UTRACE_DEATH_EVENTS and thus tracehook_report_death() can't bypass
utrace_report_death(), but we should ensure that REPORT_CALLBACKS() was
not called yet. We can check ->exit_state lockless because it must be
visible when utrace_report_death() unlocks utrace->lock.

The second check ensures we can't race with tracehook_report_death()
which checks task_utrace_flags(), so if we add _UTRACE_DEATH_EVENTS to
->utrace_flags we have to take tasklist.

In short: this double check allows to avoid tasklist when utrace_flags
already has these bits while engine->flags doesn't.

But multitracing is unlikely case, in the likely case if we add
_UTRACE_DEATH_EVENTS to engine->flags we set these bits in utrace_flags
too. I think it makes sense to consolidate these checks to make the
code a bit more understandable.

Also, add the small comment to explain the lockless exit_state check.

Signed-off-by: Oleg Nesterov <oleg@redhat.com>
Signed-off-by: Roland McGrath <roland@redhat.com>